OK, I think I'm getting the hang of how these things should be done...
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx_exp@5838 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/hash.f90
@@ -1,10 +1,10 @@
 subroutine hash(string,len,ihash)
-  use iso_c_binding, only: c_loc
+  use iso_c_binding, only: c_loc,c_size_t
   use hashing
   parameter (MASK15=32767)
 !  character*(*), target :: string
   character*1, target :: string
-  integer*8 len
+  integer(c_size_t) len
      i=nhash(c_loc(string),len,146)
      ihash=iand(i,MASK15)
   return
